requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.8.2
google-cloud-pubsub==2.23.0
cloud-sql-python-connector[pg8000]==1.12.1
